+++ previous-versions/assess/src/codatassess/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0'
-    sdk_version: str = '0.32.0'
-    gen_version: str = '2.150.0'
-    user_agent: str = 'speakeasy-sdk/python 0.32.0 2.150.0 1.0 codat-assess'
+    sdk_version: str = '0.32.1'
+    gen_version: str = '2.155.1'
+    user_agent: str = 'speakeasy-sdk/python 0.32.1 2.155.1 1.0 codat-assess'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: