+++ previous-versions/assess/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.32.0] previous-versions/assess
 ### Releases
 - [PyPI v0.32.0] https://pypi.org/project/codat-assess/0.32.0 - previous-versions/assess
+
+## 2023-10-16 01:24:26
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.1] previous-versions/assess
+### Releases
+- [PyPI v0.32.1] https://pypi.org/project/codat-assess/0.32.1 - previous-versions/assess